Fix task05 and task06
